[Deribit] Rename class + change object type

+++ xchange-deribit/src/main/java/org/knowm/xchange/deribit/v2/dto/marketdata/DeribitWithdrawalPriority.java
@@ -3,14 +3,16 @@ package org.knowm.xchange.deribit.v2.dto.marketdata;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class DeribitWithdrawalPriority {
 
-    @JsonProperty("value") public float value;
+    @JsonProperty("value") public BigDecimal value;
     @JsonProperty("name") public String name;
 
-    public float getValue() {
+    public BigDecimal getValue() {
         return value;
     }
